[Flang][MLIR][OpenMP] Use function-attached target attributes for OpenMP lowering
This patch removes the omp.target module attribute, since the information it
held on the target CPU and features is available to each function as a function
attribute. Target outlining during the MLIR to LLVM IR translation stage is
also updated, so that these attributes are passed along to the newly created
function.